+++ src/SDK.ts
@@ -2,6 +2,7 @@ import { ChannelId, LiveChatVersion } from "./Common/Enums";
 import axios, { AxiosRequestConfig, AxiosResponse } from "axios";
 import { BrowserInfo } from "./Utils/BrowserInfo";
 import Constants from "./Common/Constants";
+import { createGetChatTokenEndpoint } from "./Utils/endpointsCreators";
 import { DeviceInfo } from "./Utils/DeviceInfo";
 import FetchChatTokenResponse from "./Model/FetchChatTokenResponse";
 import IDataMaskingInfo from "./Interfaces/IDataMaskingInfo";
@@ -137,9 +138,11 @@ export default class SDK implements ISDK {
     });
     const elapsedTimeInMilliseconds = timer.milliSecondsElapsed;
     const { data } = response;
-    if (data.LiveChatVersion && data.LiveChatVersion === LiveChatVersion.V2) {
+
+    if (data.LiveChatVersion) {
       this.liveChatVersion = data.LiveChatVersion;
     }
+
     data.headers = {};
     if (response.headers && response.headers["date"]) {
       data.headers["date"] = response.headers["date"];
@@ -229,20 +232,14 @@ export default class SDK implements ISDK {
     }
 
     const headers: StringMap = Constants.defaultHeaders;
-    let requestPath = `/${OmnichannelEndpoints.LiveChatGetChatTokenPath}/${this.omnichannelConfiguration.orgId}/${this.omnichannelConfiguration.widgetId}/${requestId}`;
 
-    if (this.liveChatVersion === LiveChatVersion.V2 || (currentLiveChatVersion && currentLiveChatVersion === LiveChatVersion.V2)) {
-      requestPath = `/${OmnichannelEndpoints.LiveChatv2GetChatTokenPath}/${this.omnichannelConfiguration.orgId}/${this.omnichannelConfiguration.widgetId}/${requestId}`;
-      if (authenticatedUserToken) {
-        requestPath = `/${OmnichannelEndpoints.LiveChatv2AuthGetChatTokenPath}/${this.omnichannelConfiguration.orgId}/${this.omnichannelConfiguration.widgetId}/${requestId}`;
-        headers[OmnichannelHTTPHeaders.authenticatedUserToken] = authenticatedUserToken;
-      }
-    } else {
-      if (authenticatedUserToken) {
-        requestPath = `/${OmnichannelEndpoints.LiveChatAuthGetChatTokenPath}/${this.omnichannelConfiguration.orgId}/${this.omnichannelConfiguration.widgetId}/${requestId}`;
-        headers[OmnichannelHTTPHeaders.authenticatedUserToken] = authenticatedUserToken;
-      }
+    const endpoint = createGetChatTokenEndpoint(currentLiveChatVersion as LiveChatVersion || this.liveChatVersion, authenticatedUserToken? true: false);
+
+    if (authenticatedUserToken) {
+      headers[OmnichannelHTTPHeaders.authenticatedUserToken] = authenticatedUserToken;
     }
+
+    let requestPath = `/${endpoint}/${this.omnichannelConfiguration.orgId}/${this.omnichannelConfiguration.widgetId}/${requestId}`;
 
     if (reconnectId) {
       requestPath += `/${reconnectId}`;
@@ -1079,13 +1076,13 @@ export default class SDK implements ISDK {
 
   /**
    * Helper function for logging.
-   * 
+   *
    * @param logLevel Log level for logging.
    * @param telemetryEventType Telemetry event type in which event will be logged.
    * @param description Description of the event.
    * @param requestId Request ID
    * @param response Response
-   * @param elapsedTimeInMilliseconds Elapsed time in ms 
+   * @param elapsedTimeInMilliseconds Elapsed time in ms
    * @param requestPath Request path
    * @param method Method
    * @param error Error